+++ packages/astro/src/cli/check/index.ts
@@ -1,23 +1,21 @@
 import path from 'node:path';
 import type { Arguments } from 'yargs-parser';
-import { error, info } from '../../core/logger/core.js';
-import { createLoggingFromFlags, flagsToAstroInlineConfig } from '../flags.js';
+import { createLoggerFromFlags, flagsToAstroInlineConfig } from '../flags.js';
 import { getPackage } from '../install-package.js';
 
 export async function check(flags: Arguments) {
-	const logging = createLoggingFromFlags(flags);
+	const logger = createLoggerFromFlags(flags);
 	const getPackageOpts = { skipAsk: flags.yes || flags.y, cwd: flags.root };
 	const checkPackage = await getPackage<typeof import('@astrojs/check')>(
 		'@astrojs/check',
-		logging,
+		logger,
 		getPackageOpts,
 		['typescript']
 	);
-	const typescript = await getPackage('typescript', logging, getPackageOpts);
+	const typescript = await getPackage('typescript', logger, getPackageOpts);
 
 	if (!checkPackage || !typescript) {
-		error(
-			logging,
+		logger.error(
 			'check',
 			'The `@astrojs/check` and `typescript` packages are required for this command to work. Please manually install them into your project and try again.'
 		);
@@ -38,6 +36,6 @@ export async function check(flags: Arguments) {
 
 	const config = parseArgsAsCheckConfig(process.argv);
 
-	info(logging, 'check', `Getting diagnostics for Astro files in ${path.resolve(config.root)}...`);
+	logger.info('check', `Getting diagnostics for Astro files in ${path.resolve(config.root)}...`);
 	return await checker(config);
 }