+++ packages/astro/src/core/compile/compile.ts
@@ -45,8 +45,6 @@ export async function compile({
 			astroGlobalArgs: JSON.stringify(astroConfig.site),
 			scopedStyleStrategy: astroConfig.scopedStyleStrategy,
 			resultScopedSlot: true,
-			experimentalTransitions: astroConfig.experimental.viewTransitions,
-			experimentalPersistence: astroConfig.experimental.viewTransitions,
 			transitionsAnimationURL: 'astro/components/viewtransitions.css',
 			preprocessStyle: createStylePreprocessor({
 				filename,