+++ packages/astro/src/core/routing/validation.ts
@@ -1,7 +1,6 @@
 import type { ComponentInstance, GetStaticPathsResult, RouteData } from '../../@types/astro';
 import { AstroError, AstroErrorData } from '../errors/index.js';
-import type { LogOptions } from '../logger/core';
-import { warn } from '../logger/core.js';
+import type { Logger } from '../logger/core';
 
 const VALID_PARAM_TYPES = ['string', 'number', 'undefined'];
 
@@ -40,7 +39,7 @@ export function validateDynamicRouteModule(
 /** Throw error and log warnings for malformed getStaticPaths() response */
 export function validateGetStaticPathsResult(
 	result: GetStaticPathsResult,
-	logging: LogOptions,
+	logger: Logger,
 	route: RouteData
 ) {
 	if (!Array.isArray(result)) {
@@ -79,8 +78,7 @@ export function validateGetStaticPathsResult(
 		// TODO: Replace those with errors? They technically don't crash the build, but users might miss the warning. - erika, 2022-11-07
 		for (const [key, val] of Object.entries(pathObject.params)) {
 			if (!(typeof val === 'undefined' || typeof val === 'string' || typeof val === 'number')) {
-				warn(
-					logging,
+				logger.warn(
 					'getStaticPaths',
 					`invalid path param: ${key}. A string, number or undefined value was expected, but got \`${JSON.stringify(
 						val
@@ -88,8 +86,7 @@ export function validateGetStaticPathsResult(
 				);
 			}
 			if (typeof val === 'string' && val === '') {
-				warn(
-					logging,
+				logger.warn(
 					'getStaticPaths',
 					`invalid path param: ${key}. \`undefined\` expected for an optional param, but got empty string.`
 				);